// src/interface/tipocancionero.interface.tsx
// Generated by https://quicktype.io

export interface ResultTipoCancioneros {
    ok:             boolean;
    msg:            string;
    tipocancionero: Tipocancionero[];
}

export interface Tipocancionero {
    id:             number;
    titulo:         string;
    descripcion:    string;
    subdescripcion: string;
    numero:         string;
    logo:           string;
    id_cancionero:  number;
}